✨ Add HTTP/HTTPS server
Not working right now:
* Load Balancing
* Load Balancing strategies

Listeners:
They have been adapted to use a long-running stream of incoming sockets instead of a simple loop. The stream is then converted to a HyperAcceptor, which can be passed to hyper so it knows how to receive new sockets.

State:
At some point, we need to have the current state of the load balancer (host <-> backend mapping, round robin etc.) in the handler function of the HTTP/HTTPS server. Thus, since we have 2 servers, we had to wrap the state inside an `Arc` (Atomically Reference Counted).

We fighted over 2 hours with the borrow checker, but HUMANITY WON AGAIN! 🥳

✨ Added client to backend request forwarding

It currently only takes the first IP address of the pool.

🩹 Fixed selecting of server (random for now), still some missing parts, sorry martin �‍♂️